Update packages/overture-schema-core/src/overture/schema/core/types.py
Apply Seth's suggestion on confidence score precision.

+++ packages/overture-schema-core/src/overture/schema/core/types.py
@@ -63,7 +63,7 @@ class ConfidenceScoreConstraint(FieldConstraint):
 ConfidenceScore = NewType(
     "ConfidenceScore",
     Annotated[
-        float64,
+        float32,
         ConfidenceScoreConstraint(),
         Field(
             description="Confidence value from the source dataset, particularly relevant for ML-derived data."
